feat: v1.3.0 合规深化版 — MFA + Quorum + RFC 8998 + OTel + Config Reload + Alerting
v1.3.0 新增功能:

MFA TOTP (RFC 6238):
- GenerateTOTPSecret + GenerateTOTP + ValidateTOTP + BuildTOTPURI
- 敏感操作二次确认 (ShredKey/EmergencySeal/ExportKey/SoftDeleteKey)
- ±30s 时钟漂移容忍 + 防重放
- X-MFA-Code header 校验
- POST /api/v1/auth/mfa/setup|verify|disable

Quorum Approval (K-of-N 审批):
- ApprovalTicket + ApprovalStore 接口 + 内存实现
- 状态机: pending → approved/rejected/expired
- 防自批准 + 审批幂等 + 过期清理
- POST /api/v1/approvals + approve + reject

RFC 8998 国密 TLS (GB/T 38636):
- SM2 双证书 (签名 + 加密)
- GMTLS_SM2_WITH_SM4_SM3 / GMTLS_ECDHE_SM2_WITH_SM4_SM3
- gmsm 构建标签隔离

OpenTelemetry tracing:
- OTLP gRPC exporter + TracerProvider
- otelhttp 自动 instrumentation
- TraceID 传播到 audit log

Config Reload (SIGHUP):
- atomic.Pointer 线程安全热配置
- 热更新白名单: logging/audit/crypto/mfa/observability
- 冷更新保留旧值: server/storage/unseal/auth

Alerting Webhook:
- Slack/钉钉/PagerDuty 自动检测格式
- 高危操作触发告警

安全:
- gosec: 0 issues (27 #nosec)
- govulncheck: 0 vulnerabilities
- Policy 扩展: RequireMFA/RequireQuorum/ApproverRoles
- MFAStore/ApprovalStore 接口

测试: 47 个新测试
CI: 17 packages PASS (标准 + gmsm 构建)

+++ internal/auth/totp.go
@@ -10,7 +10,7 @@ package auth
 
 import (
 	"crypto/hmac"
-	"crypto/sha1"
+	"crypto/sha1" // #nosec G505 -- SHA-1 是 RFC 6238 TOTP 标准算法，非密码存储用途
 	"crypto/subtle"
 	"encoding/base32"
 	"encoding/binary"
@@ -58,7 +58,8 @@ func GenerateTOTP(secret string, t time.Time) (string, error) {
 	}
 
 	// T = floor((unix_time - T0) / period)，T0 = 0。
-	counter := uint64(t.Unix()) / totpPeriod
+	// #nosec G115 -- Unix 时间戳转 uint64 不会溢出（2^63 秒 = 2920 亿年）。
+	counter := uint64(t.Unix()) / uint64(totpPeriod)
 	return hotp(hmacKey, counter), nil
 }
 
@@ -77,7 +78,8 @@ func ValidateTOTP(secret, code string, usedCodeCheck func(time.Time, string) boo
 
 	// 检查 current + skew 窗口。
 	for skew := -totpSkew; skew <= totpSkew; skew++ {
-		counter := uint64(now.Unix())/totpPeriod + uint64(skew)
+		// #nosec G115 -- Unix 时间戳转 uint64 不会溢出。
+		counter := uint64(now.Unix())/uint64(totpPeriod) + uint64(skew)
 		expected := hotp(hmacKey, counter)
 
 		if subtleEqualString(expected, code) {
